+++ crates/rpc/src/method/get_state_update.rs
@@ -53,7 +53,7 @@ pub async fn get_state_update(context: RpcContext, input: Input) -> Result<Outpu
                 .pending_data
                 .get(&tx)
                 .context("Query pending data")?
-                .state_update;
+                .state_update();
 
             return Ok(Output::Pending(state_update));
         }
@@ -754,7 +754,7 @@ mod tests {
             block_id: BlockId::Pending,
         };
 
-        let expected = context.pending_data.get_unchecked().state_update;
+        let expected = context.pending_data.get_unchecked().state_update();
 
         let result = get_state_update(context, input)
             .await